Typing Attributes, Remove STTextView, Integrate Previous Features
