+++ Sources/CodeEditInputView/TextLayoutManager/TextLayoutManager+Public.swift
@@ -24,6 +24,15 @@ extension TextLayoutManager {
         lineStorage.getLine(atIndex: offset)
     }
 
+    /// Finds text line and returns it if found.
+    /// Lines are 0 indexed.
+    /// - Parameter index: The line to find.
+    /// - Returns: The text line position if any, `nil` if the index is out of bounds.
+    public func textLineForIndex(_ index: Int) -> TextLineStorage<TextLine>.TextLinePosition? {
+        guard index > 0 && index < lineStorage.count else { return nil }
+        return lineStorage.getLine(atIndex: index)
+    }
+
     public func textOffsetAtPoint(_ point: CGPoint) -> Int? {
         guard let position = lineStorage.getLine(atPosition: point.y),
               let fragmentPosition = position.data.typesetter.lineFragments.getLine(
