+++ Sources/CodeEditInputView/TextSelectionManager/TextSelectionManager+Move.swift
@@ -32,7 +32,7 @@ extension TextSelectionManager {
         }
         updateSelectionViews()
         delegate?.setNeedsDisplay()
-        NotificationCenter.default.post(Notification(name: TextSelectionManager.selectionChangedNotification))
+        NotificationCenter.default.post(Notification(name: Self.selectionChangedNotification, object: self))
     }
 
     /// Moves a single selection determined by the direction and destination provided.
