+++ Sources/CodeEditTextView/Extensions/NSRange+/NSRange+isEmpty.swift
@@ -1,0 +1,14 @@
+//
+//  NSRange+isEmpty.swift
+//  
+//
+//  Created by Khan Winter on 10/14/23.
+//
+
+import Foundation
+
+extension NSRange {
+    var isEmpty: Bool {
+        length == 0
+    }
+}
